templates/docker/sailarea/requirements.txt
torch
torchvision
torchsummary
numpy
pillow
opencv-python
websockets
flask
matplotlib
